-Added a menu and menu buttons -Changed tileset (will add more)

+++ src/entities/Entity.java
@@ -26,14 +26,9 @@ public abstract class Entity {
         g.drawRect((int)hitbox.x, (int)hitbox.y, (int)hitbox.width, (int)hitbox.height);
     }
 
-    protected void initHitbox(float x, float y, float width, float height) {
+    protected void initHitbox(float x, float y, int width, int height) {
         hitbox = new Rectangle2D.Float( x, y, width, height);
     }
-
-//    protected void updateHitbox() {
-//        hitbox.x = (int) x;
-//        hitbox.y = (int) y;
-//    }
 
     public Rectangle2D.Float getHitbox() {
         return hitbox;
